Conflicts:
	proyecto-haya/src/main/resources/optionalConfiguration/ac-plugin-haya-projectContext.xml
	rec-common/src/main/java/es/capgemini/pfs/expediente/ExpedienteManager.java
	rec-common/src/main/java/es/capgemini/pfs/interna/InternaBusinessOperation.java
	rec-web/src/main/webapp/WEB-INF/config/appMessages.properties
	recovery-coreextension-plugin/src/main/java/es/pfsgroup/plugin/recovery/mejoras/acuerdos/MEJAcuerdoManager.java
	recovery-coreextension-plugin/src/main/java/es/pfsgroup/recovery/ext/impl/acuerdo/model/EXTAcuerdo.java
	recovery-mejoras-plugin/src/main/java/es/pfsgroup/plugin/recovery/mejoras/acuerdos/controller/MEJAcuerdoController.java
	recovery-nuevoModeloBienes-plugin/src/main/resources/plugins/messages/plugin-nuevoModeloBienes-messages.properties

Former-commit-id: 848edfb87d1472d9946536775e21b9b59c412e87

+++ rec-common/src/main/java/es/capgemini/pfs/politica/PoliticaManager.java
@@ -16,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 import es.capgemini.devon.bo.BusinessOperationException;
 import es.capgemini.devon.bo.Executor;
 import es.capgemini.devon.bo.annotations.BusinessOperation;
+import es.capgemini.pfs.arquetipo.model.Arquetipo;
 import es.capgemini.pfs.comun.ComunBusinessOperation;
 import es.capgemini.pfs.configuracion.ConfiguracionBusinessOperation;
 import es.capgemini.pfs.diccionarios.DictionaryManager;
@@ -28,6 +29,7 @@ import es.capgemini.pfs.itinerario.model.DDEstadoItinerario;
 import es.capgemini.pfs.itinerario.model.DDTipoReglaVigenciaPolitica;
 import es.capgemini.pfs.itinerario.model.Estado;
 import es.capgemini.pfs.itinerario.model.ReglasVigenciaPolitica;
+import es.capgemini.pfs.persona.dao.PersonaDao;
 import es.capgemini.pfs.persona.model.Persona;
 import es.capgemini.pfs.politica.dao.CicloMarcadoPoliticaDao;
 import es.capgemini.pfs.politica.dao.PoliticaDao;
@@ -49,6 +51,9 @@ import es.capgemini.pfs.users.domain.Perfil;
 import es.capgemini.pfs.users.domain.Usuario;
 import es.capgemini.pfs.zona.model.DDZona;
 import es.capgemini.pfs.zona.model.ZonaUsuarioPerfil;
+import es.pfsgroup.commons.utils.Checks;
+import es.pfsgroup.commons.utils.dao.abm.GenericABMDao;
+import es.pfsgroup.commons.utils.dao.abm.GenericABMDao.FilterType;
 
 /**
  * Clase con los métodos de negocio relativos a los Objetivos.
@@ -69,6 +74,12 @@ public class PoliticaManager {
 
     @Autowired
     private DictionaryManager dictionaryManager;
+    
+    @Autowired
+	private GenericABMDao genericDao;
+    
+    @Autowired
+    private PersonaDao personaDao;
 
     private final Log logger = LogFactory.getLog(getClass());
 
@@ -239,6 +250,34 @@ public class PoliticaManager {
     public List<DDTipoPolitica> getTipoPoliticaList() {
         return (List<DDTipoPolitica>) executor.execute(ComunBusinessOperation.BO_DICTIONARY_GET_LIST, DDTipoPolitica.class.getName());
     }
+    
+    @BusinessOperation(InternaBusinessOperation.BO_POL_MGR_GET_TIPO_POLITICA_PERSONA_LIST)
+    public List<DDTipoPolitica> getTipoPoliticaPersonaList(Long idPersona) {
+    	//Si no podemos obtener la politica de la entidad para la persona
+    	//Devolvemos todos los tipos de politicas
+    	if (idPersona==null)
+    		return this.getTipoPoliticaList();
+    	
+    	Persona persona = personaDao.get(idPersona);
+    	if (persona==null)
+    		return this.getTipoPoliticaList();
+    	
+    	if (persona.getPoliticaEntidad()==null)
+    		return this.getTipoPoliticaList();
+    	
+    	//Si tenemos informado para la persona un tipo de politica de la entidad
+    	//devolvemos el diccionario filtrado por los tipos que le corresponden
+    	List<DDTipoPolitica> politicas = (List<DDTipoPolitica>)genericDao.getList(DDTipoPolitica.class,
+    				genericDao.createFilter(FilterType.EQUALS, "politicaEntidad.id", persona.getPoliticaEntidad().getId()),
+    				genericDao.createFilter(FilterType.EQUALS, "auditoria.borrado", false));
+    	
+    	//Si no hemos obtenido ninguna politica de la traducción, devolvemos todas
+    	if (Checks.estaVacio(politicas)) {
+    		return this.getTipoPoliticaList();
+    	} else {
+    		return politicas;
+    	}
+    }
 
     /**
      * Recupera los permisos del usuario conectado sobre los objetivos de la política (si es gestor o supervisor).
@@ -362,7 +401,6 @@ public class PoliticaManager {
             DDEstadoItinerarioPolitica estadoItinerarioPolitica = (DDEstadoItinerarioPolitica) executor
                     .execute(ComunBusinessOperation.BO_DICTIONARY_GET_BY_CODE, DDEstadoItinerarioPolitica.class,
                             DDEstadoItinerarioPolitica.ESTADO_PREPOLITICA);
-            DDTipoPolitica politicaPersona = persona.getPrepolitica();
 
             Politica prepolitica = new Politica();
 
@@ -373,9 +411,33 @@ public class PoliticaManager {
             prepolitica.setEstadoItinerarioPolitica(estadoItinerarioPolitica);
             prepolitica.setCicloMarcadoPolitica(cicloMarcado);
 
+            /**DDTipoPolitica politicaPersona = persona.getPrepolitica();
             if (politicaPersona != null) {
                 prepolitica.setTipoPolitica(politicaPersona);
+            }**/
+            //Nueva forma de obtener el tipo de política
+            
+            
+            DDTipoPolitica tipoPolitica = null;
+            Arquetipo arquetipo = null;
+            
+            //Si el expediente es automático se obtiene el arquetipo de la persona, de la tabla ARR
+            if (!expediente.isManual()) {
+            	arquetipo = (Arquetipo) executor.execute(ConfiguracionBusinessOperation.BO_ARQ_MGR_GET_RECUPERACION_BY_PERSONA, persona.getId());
+            } else {
+            	//Si no el arquetipo se preselecciona en la creación del expedietne manual
+            	arquetipo = expediente.getArquetipo();
             }
+            	
+            
+            if( arquetipo!=null && arquetipo.getItinerario()!=null){
+            	tipoPolitica = arquetipo.getItinerario().getPrePolitica();
+            }
+            
+            if (tipoPolitica!=null) {
+                prepolitica.setTipoPolitica(tipoPolitica);
+            }
+            
             politicaDao.save(prepolitica);
 
             // ********************************************************************** //
@@ -410,9 +472,9 @@ public class PoliticaManager {
                 politicaCE.setZonaGestor(zonaGestor);
                 politicaCE.setZonaSupervisor(zonaSupervisor);
             }
-
-            if (politicaPersona != null) {
-                politicaCE.setTipoPolitica(politicaPersona);
+            //Nueva forma de obtener el tipo de política para CE
+            if (tipoPolitica!=null) {
+                politicaCE.setTipoPolitica(tipoPolitica);
             }
 
             politicaDao.save(politicaCE);
@@ -430,7 +492,11 @@ public class PoliticaManager {
                     .execute(ConfiguracionBusinessOperation.BO_ZONA_MGR_EXISTE_PERFIL_ZONA, zonaBusqueda.getId(), idPerfil);
             if (existe) return zonaBusqueda;
 
-            zonaBusqueda = zonaBusqueda.getZonaPadre();
+            if (!Checks.esNulo(zonaBusqueda.getZonaPadre()) && zonaBusqueda == zonaBusqueda.getZonaPadre()) {
+            	zonaBusqueda = null;
+            } else {
+            	zonaBusqueda = zonaBusqueda.getZonaPadre();
+            }
         }
 
         throw new GenericRollbackException("No se encuentra correspondencia para zona-perfil: " + zonaExpediente.getId() + "-" + perfilOrigen.getId());
@@ -1073,16 +1139,17 @@ public class PoliticaManager {
 
         for (CicloMarcadoPolitica cmp : listadoCiclos) {
             List<Politica> listadoPoliticas = cmp.getPoliticas();
-
-            Politica politicaBorrar = listadoPoliticas.get(listadoPoliticas.size() - 1);
-            Politica politicaProponer = listadoPoliticas.get(listadoPoliticas.size() - 2);
-
-            DDEstadoPolitica estadoPropuesta = (DDEstadoPolitica) dictionaryManager.getByCode(DDEstadoPolitica.class,
-                    DDEstadoPolitica.ESTADO_PROPUESTA);
-            politicaProponer.setEstadoPolitica(estadoPropuesta);
-
-            politicaDao.delete(politicaBorrar);
-            politicaDao.update(politicaProponer);
+            if (listadoPoliticas.size()>1) {
+	            Politica politicaBorrar = listadoPoliticas.get(listadoPoliticas.size() - 1);
+	            Politica politicaProponer = listadoPoliticas.get(listadoPoliticas.size() - 2);
+	
+	            DDEstadoPolitica estadoPropuesta = (DDEstadoPolitica) dictionaryManager.getByCode(DDEstadoPolitica.class,
+	                    DDEstadoPolitica.ESTADO_PROPUESTA);
+	            politicaProponer.setEstadoPolitica(estadoPropuesta);
+	
+	            politicaDao.delete(politicaBorrar);
+	            politicaDao.update(politicaProponer);
+            }
         }
     }
 
@@ -1110,6 +1177,15 @@ public class PoliticaManager {
         //Marcamos la política como histórica
         politica.setEstadoPolitica((DDEstadoPolitica) dictionaryManager.getByCode(DDEstadoPolitica.class, DDEstadoPolitica.ESTADO_HISTORICA));
         politicaDao.save(politica);
+    }
+    
+    /**
+     * Devuelve los tipos de política no marcados como borrado
+     * @return List<DDTipoPolitica> Lista de tipos de política
+     */
+    @BusinessOperation(InternaBusinessOperation.BO_POL_MGR_GET_TIPOS_POLITICA)
+    public List<DDTipoPolitica> getTiposPolitica() {
+        return genericDao.getList(DDTipoPolitica.class, genericDao.createFilter(FilterType.EQUALS, "auditoria.borrado", false));
     }
 
 }
